requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
openpyxl
